Made the temporary functions that are now called by packets, public so they have access to them.
Removed Most of the process_packet_contents as it is mostly all handled by the PacketHandler now

The encryption seed opcode will remain here, all others that remain will be moved to actual packets when those packets are created.

Added an getChatInterface() method so that the chat interface can be accessed from packets.  This should be moved to another file at some point.  Maybe a master interface object that deals with all of them, or make ChatBox a singleton or something.